+++ persistent/Database/Persist/Class/PersistEntity.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
@@ -33,6 +34,9 @@ module Database.Persist.Class.PersistEntity
     , toPersistValueEnum, fromPersistValueEnum
       -- * Support for @OverloadedLabels@ with 'EntityField'
     , SymbolToField (..)
+    , -- * Safety check for inserts
+      SafeToInsert
+    , SafeToInsertErrorMessage
     ) where
 
 import Data.Functor.Constant
@@ -546,3 +550,37 @@ class SymbolToField (sym :: Symbol) rec typ | sym rec -> typ where
 -- @since 2.11.0.0
 instance SymbolToField sym rec typ => IsLabel sym (EntityField rec typ) where
     fromLabel = symbolToField @sym
+
+-- | A type class which is used to witness that a type is safe to insert into
+-- the database without providing a primary key.
+--
+-- The @TemplateHaskell@ function 'mkPersist' will generate instances of this
+-- class for any entity that it works on. If the entity has a default primary
+-- key, then it provides a regular instance. If the entity has a @Primary@
+-- natural key, then this works fine. But if the entity has an @Id@ column with
+-- no @default=@, then this does a 'TypeError' and forces the user to use
+-- 'insertKey'.
+--
+-- @since 2.14.0.0
+class SafeToInsert a where
+
+type SafeToInsertErrorMessage a
+    = 'Text "The PersistEntity " ':<>: ShowType a ':<>: 'Text " does not have a default primary key."
+    ':$$: 'Text "This means that 'insert' will fail with a database error."
+    ':$$: 'Text "Please  provide a default= clause inthe entity definition,"
+    ':$$: 'Text "or use 'insertKey' instead to provide one."
+
+instance (TypeError (FunctionErrorMessage a b)) => SafeToInsert (a -> b)
+
+type FunctionErrorMessage a b =
+    'Text "Uh oh! It looks like you are trying to insert a function into the database."
+    ':$$: 'Text "Argument: " ':<>: 'ShowType a
+    ':$$: 'Text "Result:   " ':<>: 'ShowType b
+    ':$$: 'Text "You probably need to add more arguments to an Entity construction."
+
+type EntityErrorMessage a =
+    'Text "It looks like you're trying to `insert` an `Entity " ':<>: 'ShowType a ':<>: 'Text "` directly."
+    ':$$: 'Text "You want `insertKey` instead. As an example:"
+    ':$$: 'Text "    insertKey (entityKey ent) (entityVal ent)"
+
+instance TypeError (EntityErrorMessage a) => SafeToInsert (Entity a)