+++ persistent/ChangeLog.md
@@ -4,6 +4,10 @@
 
 * [#1343](https://github.com/yesodweb/persistent/pull/1343)
     * Implement Type Literal based field definitions
+* [#1387](https://github.com/yesodweb/persistent/pull/1387)
+    * Better UX with `insert`. We now report a type error when you try to
+      `insert` an `Entity` or a function, and we also forbid `insert`ing if the
+      database would throw an error missing a primary key.
 * [#1383](https://github.com/yesodweb/persistent/pull/1383)
     * Primary keys have a `NonEmpty` of fields, not a `[]` of fields.
     * A `Primary` key on an entity now creates a `Unique` constructror for that